requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
langchain
langchain-openai
langchain-community
faiss-cpu==1.8.0
fastapi==0.115.0
uvicorn==0.30.6
python-dotenv==1.0.1
pydantic==2.7.3
